# docs/swagger.yaml
components:
  schemas:
...
    handlers.loginDTO:
      properties:
        password:
          maxLength: 64
          minLength: 3
          type: string
        username:
          maxLength: 64
          minLength: 3
          type: string
      required:
      - password
      - username
      type: object
    handlers.registerDTO:
      properties:
        password:
          maxLength: 64
          minLength: 3
          type: string
        username:
          maxLength: 64
          minLength: 3
          type: string
      required:
      - password
      - username
      type: object
...
    models.Error:
      properties:
        error:
          type: string
        message:
          type: string
      type: object
...
    sqlc.AuthInsertUserRow:
      properties:
        displayname:
          type: string
        username:
          type: string
      type: object
    sqlc.AuthSelectUserRow:
      properties:
        password:
          type: string
        username:
          type: string
      type: object
...
    sqlc.RecitationFileUpdateRecitationFileParams:
      properties:
        has_timings:
          type: boolean
        lafzize_processing:
          type: boolean
        reciter:
          type: string
        slug:
          type: string
        verse_key:
          type: string
      type: object
    sqlc.Session:
      properties:
        csrf_token:
          type: string
        session_token:
          type: string
        username:
          type: string
      type: object
...
info:
  title: tilawah-hub
  version: 0.0.1
openapi: 3.1.0
paths:
  /lafzize/{slug}/{verse_key}:
    post:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      - description: Recitation slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Verse key of recitation
        in: path
        name: verse_key
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.RecitationFileUpdateRecitationFileParams'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
      tags:
      - lafzize
  /login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.loginDTO'
        description: Login
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.AuthSelectUserRow'
          description: OK
          headers:
            "CSRF\tToken":
              schema:
                type: string
            "Session\tToken":
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - Auth
  /logout:
    post:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.Session'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - Auth
...
  /register:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.registerDTO'
        description: Register
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.AuthInsertUserRow'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - Auth
...