# internal/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package sqlc

type Recitation struct {
	Slug    string `json:"slug"`
	Name    string `json:"name"`
	Reciter string `json:"reciter"`
}

type RecitationFile struct {
	Reciter           string `json:"reciter"`
	Slug              string `json:"slug"`
	VerseKey          string `json:"verse_key"`
	HasTimings        bool   `json:"has_timings"`
	LafzizeProcessing bool   `json:"lafzize_processing"`
}

type Session struct {
	SessionToken string `json:"session_token"`
	CsrfToken    string `json:"csrf_token"`
	Username     string `json:"username"`
}

type User struct {
	Username    string `json:"username"`
	Password    string `json:"password"`
	Displayname string `json:"displayname"`
}